+++ integer/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `UBig::ones`.
 - Add `IBig::as_ubig`.
+- Add `UBig::from_chunks` and `UBig::to_chunks`.
 - Implement `TryFrom<UBig>` and `TryFrom<IBig>` for `f32`/`f64`.
 - Implement `IBig::{from_le_bytes, from_be_bytes}` and `IBig::{to_le_bytes, to_be_bytes}`.
 - The alterative `Debug` output of `UBig` and `IBig` will include `(digits: x, bits: y)` instead of `(x digits, y bits)`.